Refactor refreshable
